pyproject.toml
[project]
name = "hirawilliott"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.8",
    "google-cloud-texttospeech>=2.24.0",
    "hypermedia>=5.3.3",
]
